Deleting old pagespeed and metric data.

+++ deleteolddata.php
@@ -3,6 +3,18 @@ require_once('global.php');
 
 if ($oldDataInterval > 0)
 {
+	# deleting old data for custom metrics 
+	$query = sprintf("DELETE FROM metric WHERE timestamp < DATE_SUB(now(), INTERVAL '%s' DAY)",
+		mysql_real_escape_string($oldDataInterval)
+	);
+
+	$result = mysql_query($query);
+
+	if (!$result) {
+		error_log(mysql_error());
+		exit;
+	}
+
 	# deleting old data for yslow v2
 	$query = sprintf("DELETE FROM yslow2 WHERE timestamp < DATE_SUB(now(), INTERVAL '%s' DAY)",
 		mysql_real_escape_string($oldDataInterval)
@@ -15,8 +27,20 @@ if ($oldDataInterval > 0)
 		exit;
 	}
 
+	# deleting old data for pagespeed
+	$query = sprintf("DELETE FROM pagespeed WHERE timestamp < DATE_SUB(now(), INTERVAL '%s' DAY)",
+		mysql_real_escape_string($oldDataInterval)
+	);
+
+	$result = mysql_query($query);
+
+	if (!$result) {
+		error_log(mysql_error());
+		exit;
+	}
+
 	# updating latest values for the URL
-	$query = 'DELETE urls FROM urls LEFT JOIN (SELECT DISTINCT url_id FROM yslow2) AS y ON urls.id = y.url_id WHERE y.url_id IS NULL';
+	$query = 'DELETE urls FROM urls LEFT JOIN (SELECT DISTINCT url_id FROM yslow2) AS y ON urls.id = y.url_id LEFT JOIN (SELECT DISTINCT url_id FROM pagespeed) AS p ON urls.id = p.url_id LEFT JOIN (SELECT DISTINCT url_id FROM metric) AS m ON urls.id = m.url_id WHERE y.url_id IS NULL AND p.url_id IS NULL AND m.url_id IS NULL';
 
 	$result = mysql_query($query);
